[Gold V] Title: 피타고라스 정리의 증명, Time: 348 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/33466. 피타고라스 정리의 증명/피타고라스 정리의 증명.cc
@@ -14,7 +14,7 @@ int main() {
 	int T; cin >> T;
 
 	while (T--) {
-		int n; cin >> n;
+		ll n; cin >> n;
 		
 		if (n & 1) {
 			cout << n * 2 - 1 << '\n';
